deps: set typescript version
