+++ .projenrc.ts
@@ -4,6 +4,7 @@ const project = new awscdk.AwsCdkConstructLibrary({
   author: 'yicr',
   authorAddress: 'yicr@users.noreply.github.com',
   cdkVersion: '2.43.0',
+  typescriptVersion: '5.1.x',
   defaultReleaseBranch: 'main',
   projenrcTs: true,
   name: '@gammarer/aws-secure-frontend-web-app-cloudfront-distribution',
