+++ package.json
@@ -58,7 +58,7 @@
     "standard-version": "^9",
     "ts-jest": "^27",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.5"
+    "typescript": "5.1.x"
   },
   "peerDependencies": {
     "aws-cdk-lib": "^2.43.0",
